Move git activity stats data extraction to git module
